[#14147] Update documentation to reflect new Protostream based GlobalMarshaller
- Consistently use ProtoStream vs Protostream throughout documentation
