--- documentation/src/main/asciidoc/topics/code_examples/ABCMarshallingExternalizer.java
@@ -1,15 +1,0 @@
-public static class ABCMarshallingExternalizer implements AdvancedExternalizer<ABCMarshalling> {
-   @Override
-   public void writeObject(ObjectOutput output, ABCMarshalling object) throws IOException {
-      output.writeObject(object.getMap());
-   }
-
-   @Override
-   public ABCMarshalling readObject(ObjectInput input) throws IOException, ClassNotFoundException {
-      ABCMarshalling hi = new ABCMarshalling();
-      hi.setMap((ConcurrentHashMap<Long, Long>) input.readObject());
-      return hi;
-   }
-
-   ...
-}
