--- documentation/src/main/asciidoc/topics/code_examples/ABCMarshallingExternalizerWrong.java
@@ -1,17 +1,0 @@
-public static class ABCMarshallingExternalizer implements AdvancedExternalizer<ABCMarshalling> {
-   @Override
-   public void writeObject(ObjectOutput output, ABCMarshalling object) throws IOException {
-      MapExternalizer ma = new MapExternalizer();
-      ma.writeObject(output, object.getMap());
-   }
-
-   @Override
-   public ABCMarshalling readObject(ObjectInput input) throws IOException, ClassNotFoundException {
-      ABCMarshalling hi = new ABCMarshalling();
-      MapExternalizer ma = new MapExternalizer();
-      hi.setMap((ConcurrentHashMap<Long, Long>) ma.readObject(input));
-      return hi;
-   }
-
-   ...
-}
