--- documentation/src/main/asciidoc/topics/code_examples/AdvancedExternalizerAddBookExternalizer.java
@@ -1,2 +1,0 @@
-GlobalConfigurationBuilder builder = ...
-builder.serialization().addAdvancedExternalizer(new Book.BookExternalizer());
