--- documentation/src/main/asciidoc/topics/code_examples/ConfigExternalizers.java
@@ -1,5 +1,0 @@
-GlobalConfiguration globalConfig = new GlobalConfigurationBuilder()
-  .serialization()
-    .addAdvancedExternalizer(998, new PersonExternalizer())
-    .addAdvancedExternalizer(999, new AddressExternalizer())
-  .build();
