--- documentation/src/main/asciidoc/topics/code_examples/ExternalizerGetTypeClasses.java
@@ -1,14 +1,0 @@
-import org.infinispan.commons.util.Util;
-...
-@Override
-public Set<Class<? extends ReplicableCommand>> getTypeClasses() {
-  return Util.asSet(LockControlCommand.class, RehashControlCommand.class,
-      StateTransferControlCommand.class, GetKeyValueCommand.class,
-      ClusteredGetCommand.class,
-      SingleRpcCommand.class, CommitCommand.class,
-      PrepareCommand.class, RollbackCommand.class,
-      ClearCommand.class, EvictCommand.class,
-      InvalidateCommand.class, InvalidateL1Command.class,
-      PutKeyValueCommand.class, PutMapCommand.class,
-      RemoveCommand.class, ReplaceCommand.class);
-}
