--- documentation/src/main/asciidoc/topics/code_examples/ExternalizerRegisterId.java
@@ -1,3 +1,0 @@
-GlobalConfigurationBuilder builder = ...
-builder.serialization()
-   .addAdvancedExternalizer(123, new Person.PersonExternalizer());
