--- documentation/src/main/asciidoc/topics/code_examples/ExternalizerRegisterMultiple.java
@@ -1,3 +1,0 @@
-builder.serialization()
-   .addAdvancedExternalizer(new Person.PersonExternalizer(),
-                            new Address.AddressExternalizer());
