--- documentation/src/main/asciidoc/topics/code_examples/ExternalizerRegisterPerson.java
@@ -1,3 +1,0 @@
-GlobalConfigurationBuilder builder = ...
-builder.serialization()
-   .addAdvancedExternalizer(new Person.PersonExternalizer());
