+++ documentation/src/main/asciidoc/topics/code_examples/GenericJBossMarshaller.java
