+++ documentation/src/main/asciidoc/topics/functional_api.adoc
@@ -466,16 +466,15 @@ include::code_examples/WriteOnlyMapWriteFuture.java[]
 
 without doing the cast described above.
 
-A more economical way to marshall a function is to provide an {brandname}
-link:../../apidocs/org/infinispan/commons/marshall/Externalizer.html[`Externalizer`] for it:
+A more economical way to marshall a function is to utilise ProtoStream annotations:
 
 [source,java]
 ----
-include::code_examples/WriteOnlyMapExternalizer.java[]
+include::code_examples/WriteOnlyMapProtoStream.java[]
 ----
 
 To help users take advantage of the tiny payloads generated by
-`Externalizer`-based functions, the functional API comes with a helper
+ProtoStream annotation functions, the functional API comes with a helper
 class called
 link:../../apidocs/org/infinispan/commons/marshall/MarshallableFunctions.html[`org.infinispan.commons.marshall.MarshallableFunctions`]
 which provides marshallable functions for some of the most commonly user
