--- documentation/src/main/asciidoc/topics/xml/advanced_externalizer_book_example.xml
@@ -1,5 +1,0 @@
-<cache-container>
-   <serialization>
-      <advanced-externalizer class="Book$BookExternalizer"/>
-   </serialization>
-</cache-container>
